angler: Use init.chopped64.rc conditionally
If "Chopped64" exists in the defconfig, then clearly we're on the chopped64 branch, and want to use the Chopped64 kernel. So only copy that if we're using that kernel.

init.chopped64.rc is modified from init.flash.rc of FlashROM/FlashKernel, so major credits and thanks to @nathanchance for the initial version.

Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>